requirements.txt
Flask
tabula-py
pandas
requests
python-dotenv
